progress integrating wasm
I started integrating WASM into the actual webpack build environment,
but I decided that I need to instead just completely scaffold this
project differently. CRA uses too old a version of webpack to be dealing
with such new technology. So we're going to customize things a bit in
the next version of `thors-react-app`, then merge that back to here.